Renaming namespaces to folder structure
