+++ TimberAPI/AssetLoaderSystem/AssetSystem/AssetRepository.cs
@@ -1,8 +1,8 @@
 using System.Collections.Generic;
 using System.Linq;
-using TimberbornAPI.AssetLoader.Exceptions;
+using TimberbornAPI.AssetLoaderSystem.Exceptions;
 
-namespace TimberbornAPI.AssetLoader.AssetSystem
+namespace TimberbornAPI.AssetLoaderSystem.AssetSystem
 {
     internal class AssetRepository : IAssetRepository
     {
