+++ TimberAPI/DependencySystem/IDependencyRegistry.cs
