+++ TimberAPI/LocalizationSystem/ILocalization.cs
