+++ TimberAPI/LocalizationSystem/LocalizationRecord.cs
