+++ TimberAPI/UIBuilderSystem/CustomElements/LocalizableSliderInt.cs
