+++ TimberAPIExample/Examples/EntityActionExample/TriggerActionExample.cs
@@ -1,6 +1,6 @@
 using Timberborn.Buildings;
 using Timberborn.ConstructibleSystem;
-using TimberbornAPI.EntityInstantiatorSystem;
+using TimberbornAPI.EntityActionSystem;
 using UnityEngine;
 
 namespace TimberAPIExample.Examples.EntityActionExample
